[pallas] support ordered jagged carry under fori_loop
Data-dependent tiles default to fori_loop, but ordered jagged carry only ran
under emit_pipeline. A row map with an unaligned group boundary could
therefore not lower safely through the default path.

With offsets [0, 3, 16] and a 16-row block, both groups touch the aligned
window [0, 16). Carry must preserve the first group when the second group
stores.

Extend the shared InnerLoopWindow to fori_loop. Use its aligned grid, begin,
step, and two-sided masks; annotate DMA offsets with their provable sublane
alignment; and write full carried blocks after the boundary fold. Dimensions
that do not carry retain their clamped stores.

The default configuration can now lower jagged BMM and elementwise row maps
through fori_loop. Both streaming lowerings also apply the same unsafe-write
rejections.

Compatibility:
A bf16 y[:, tile, :] load followed by a major-dimension store is refused
here rather than lowered. The first commit of this series already stopped it
compiling, by declining to annotate an offset it could not prove aligned;
now that fori_loop rounds the window, the write on that window is rejected
explicitly instead. Such a kernel could previously appear correct whenever
every group began on a sublane boundary, but corrupt preceding-group rows at
unaligned boundaries.

Read-only aligned reductions remain rejected, so the flex attention example
still does not lower. The next commit allows them and restores it.

stack-info: PR: https://github.com/pytorch/helion/pull/3373, branch: cota/stack/12

+++ test/test_pallas.py
@@ -5980,7 +5980,10 @@ class TestPallas(TestCase):
         H, D = 4, 16
         offsets = torch.tensor([0, 10], device=DEVICE, dtype=torch.int32)
         torch.manual_seed(0)
-        y = torch.randn(H, 10, D, device=DEVICE, dtype=torch.bfloat16)
+        # Keep this mask-placement test directly addressable. A bf16 load
+        # would need a rounded-down window, and the major-dim output store
+        # cannot carry its head rows, so this commit rejects that shape.
+        y = torch.randn(H, 10, D, device=DEVICE, dtype=torch.float32)
 
         code, out = code_and_output(
             opposite,
